solution/requirements.txt
pandas
mysqlclient
mysql-connector-python
SQLAlchemy